+++ tests/test_amend.py
@@ -26,7 +26,7 @@ from edumatcher.models.order import (
 
 def _limit(
     side: Side,
-    price: float,
+    price: int,
     qty: int = 100,
     gateway_id: str = "GW1",
 ) -> Order:
@@ -42,7 +42,7 @@ def _limit(
 
 def _iceberg(
     side: Side,
-    price: float,
+    price: int,
     qty: int = 300,
     visible_qty: int = 100,
 ) -> Order:
